Added blogs and concept of root directory
Added my first blog to the website. Sampling theorem for the bandpass
signals. Also removed the sidebar code from every document and added
the concept of root folder to reuse the code. Insted of adding the
sidebar code to every file, I am adding onliner javascript to every
page that just states the current ID to highlight in the sidebar.
TODO: Add table of content and title to the blogs entries.